ml/ml-dev/ml-libs/analytics/google-ads/google-ads-poc/requirements.txt
google-ads
pyyaml
python-dotenv